+++ src/screens/Ticketing/Purchase/utils.ts
@@ -1,4 +1,6 @@
 import {PreassignedFareProduct} from '@atb/reference-data/types';
+import { PaymentType } from '@atb/tickets/types';
+import { parseISO } from 'date-fns';
 
 export type PurchaseFlow = {
   /**
@@ -29,3 +31,28 @@ export const getPurchaseFlow = (
     };
   }
 };
+
+export function getExpireDate(iso: string): string {
+  let date = parseISO(iso);
+  let year = date.getFullYear();
+  let month = date.getMonth();
+  if (month === 0) {
+    month = 12;
+    year--;
+  }
+  return `${month < 10 ? '0' + month : month}/${year.toString().slice(2, 4)}`;
+}
+
+export function getPaymentTypeName(paymentType: PaymentType) {
+  switch (paymentType) {
+    case PaymentType.Visa:
+      return 'Visa';
+    case PaymentType.Mastercard:
+      return 'MasterCard';
+    case PaymentType.Vipps:
+      return 'Vipps';
+    default:
+      return '';
+  }
+}
+
